Controller y vistas Producto modificadas

+++ Models/Presupuesto.cs
@@ -8,13 +8,13 @@ namespace tl2_tp8_2025_camip1.Models
         public List<PresupuestoDetalle> ListaDetalles { get; set; }
 
         //METODOS
-        public int MontoPresupuesto()
+        public decimal MontoPresupuesto()
         {
             return ListaDetalles.Sum(d => d.Producto.Precio * d.Cantidad);
         }
-        public double MontoPresupuestoConIva() //considerar iva 21
+        public decimal MontoPresupuestoConIva() //considerar iva 21
         {
-            int presupuesto = (int)(MontoPresupuesto() * 1.21);
+            decimal presupuesto = MontoPresupuesto() * (decimal)1.21;
             return presupuesto;
         }
         public int CantidadProductos() //contar total de productos (sumador de todas las cantidades del detalle)
